Create&Read para postagens

+++ web-hola-mundo/src/main/java/com/webii/webholamundo/repository/web2Repository.java
@@ -3,5 +3,7 @@ package com.webii.webholamundo.repository;
 import com.webii.webholamundo.model.Postagens;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface web2Repository extends JpaRepository <Postagens, Long> {
+import java.util.List;
+
+public interface web2Repository extends JpaRepository <Postagens, Integer> {
 }
